Add a metrics/mod.rs, push additional kinesis metrics
