+++ src/dataflow/lib.rs
@@ -11,6 +11,7 @@
 
 mod arrangement;
 mod decode;
+mod metrics;
 mod operator;
 mod render;
 mod sink;
